refactor MainCalculatorWindow to separate class for testing
